requirements.txt
Flask==2.3.3
playwright==1.40.0
python-dotenv==1.0.0
gunicorn==21.2.0
flask-cors==4.0.0
loguru==0.7.2 